feat(build): build aur pkgs during build (#51)

+++ build_files/mainsys/03-subsystem.sh
@@ -11,16 +11,24 @@ wget -q https://github.com/1player/host-spawn/releases/download/v1.6.2/host-spaw
 chmod +x /mkosi/mkosi.extra/usr/libexec/host-spawn
 
 # build arch rootfs
-mkosi build --directory="/mkosi" --environment="IMAGE_VARIANT=$IMAGE_VARIANT"
+if ! mkosi build --directory="/mkosi" --environment="IMAGE_VARIANT=$IMAGE_VARIANT" >/tmp/mkosi.log 2>&1; then
+    tail -n 200 /tmp/mkosi.log
+    exit 1
+fi
 
-# install extra pkgs into rootfs
-pacman -U --root /mkosi/output/image --noconfirm /packages/subsys/* >/dev/null
+# install extra pkgs
+if ! su - builder -c "xargs -a /mkosi/conf/01-aur-pkgs yay -S --noconfirm --needed --root=/mkosi/output/image/" >/tmp/yay.log 2>&1; then
+    tail -n 200 /tmp/yay.log
+    exit 1
+fi
 
 # compress rootfs
 mkdir -p /usr/lib/subsystem/rootfs
 mkfs.erofs -zzstd,5 -C 65536 -E all-fragments,dedupe,fragdedupe=inode -L rootfs /usr/lib/subsystem/rootfs/rootfs.dsk /mkosi/output/image >/dev/null
 
 # cleanup
-rm -rf /mkosi
+pacman -Rns --noconfirm yay-bin
+userdel builder
+rm -rf /mkosi /tmp/mkosi.log /tmp/yay.log
 
 echo "::endgroup::"
